#66 - test: 테스트코드 수정
