+++ src/test/java/com/example/peep/service/QuestionServiceTest.java
@@ -9,10 +9,7 @@ import com.example.peep.domain.mapping.CommunityQuestion;
 import com.example.peep.domain.mapping.StudentCommunity;
 import com.example.peep.domain.mapping.StudentCommunityQuestion;
 import com.example.peep.dto.CommunityQuestionDto;
-import com.example.peep.repository.CommunityQuestionRepository;
-import com.example.peep.repository.StudentCommunityQuestionRepository;
-import com.example.peep.repository.StudentCommunityRepository;
-import com.example.peep.repository.StudentRepository;
+import com.example.peep.repository.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -35,6 +32,8 @@ class QuestionServiceTest {
     @Mock private StudentRepository studentRepository;
     @Mock private StudentCommunityRepository studentCommunityRepository;
     @Mock private CommunityQuestionRepository communityQuestionRepository;
+    @Mock private CommunityRepository communityRepository;
+    @Mock private QuestionRepository questionRepository;
     @Mock private StudentCommunityQuestionRepository studentCommunityQuestionRepository;
 
     @InjectMocks
@@ -110,4 +109,31 @@ class QuestionServiceTest {
 
     }
 
+    @DisplayName("updateRandomQuestion - 커뮤니티에 시간이 지나면 자동으로 질문 추가")
+    @Test
+    void given_whenUpdateRandomQuestion_thenUpdateQuestion() {
+        //Given
+        School school = School.of("교육청", "학교");
+        Community community1 = Community.of(school, 3,1);
+        Community community2 = Community.of(school, 3,1);
+        Community community3 = Community.of(school, 3,1);
+        List<Community> communities = new ArrayList<>();
+        communities.add(community1);
+        communities.add(community2);
+        communities.add(community3);
+        Question q1 = Question.of("question1");
+        Question q2 = Question.of("question1");
+        Question q3 = Question.of("question1");
+        List<Question> qList = new ArrayList<>();
+        qList.add(q1);
+        qList.add(q2);
+        qList.add(q3);
+        given(communityRepository.findAll()).willReturn(communities);
+        given(questionRepository.findRandomQuestions()).willReturn(qList);
+        //When
+        questionService.updateRandomCommunityQuestion();
+        //Then
+        then(communityQuestionRepository).should().saveAll(anyList());
+
+    }
 }
